refactor(wordcloud): introduce the wc core module
