+++ mwc/bots/cli.py
@@ -5,7 +5,7 @@ from mwc.core.cfg import load_config
 
 from mwc.core.helpers import get_stop_words
 from mwc.core.db.queries import get_next_movie
-from mwc.core.wordcloud import WordCloud
+from mwc.core.wc.wordcloud import WordCloud
 from .twitter import TwitterClient
 
 log = logging.getLogger(__name__)
